Add support for getting disk info on iOS
