+++ src/apple/ffi.rs
@@ -1,10 +1,30 @@
 // Take a look at the license at the top of the repository in the LICENSE file.
 
+use core_foundation_sys::{
+    array::CFArrayRef, dictionary::CFDictionaryRef, error::CFErrorRef, string::CFStringRef,
+    url::CFURLRef,
+};
+
 // Reexport items defined in either macos or ios ffi module.
 pub use crate::sys::inner::ffi::*;
 
 #[link(name = "CoreFoundation", kind = "framework")]
-extern "C" {}
+extern "C" {
+    pub fn CFURLCopyResourcePropertiesForKeys(
+        url: CFURLRef,
+        keys: CFArrayRef,
+        error: *mut CFErrorRef,
+    ) -> CFDictionaryRef;
+
+    pub static kCFURLVolumeIsEjectableKey: CFStringRef;
+    pub static kCFURLVolumeIsRemovableKey: CFStringRef;
+    pub static kCFURLVolumeAvailableCapacityKey: CFStringRef;
+    pub static kCFURLVolumeAvailableCapacityForImportantUsageKey: CFStringRef;
+    pub static kCFURLVolumeTotalCapacityKey: CFStringRef;
+    pub static kCFURLVolumeNameKey: CFStringRef;
+    pub static kCFURLVolumeIsLocalKey: CFStringRef;
+    pub static kCFURLVolumeIsBrowsableKey: CFStringRef;
+}
 
 #[cfg_attr(feature = "debug", derive(Eq, Hash, PartialEq))]
 #[derive(Clone)]
